Pass `CycleApprox`s into `triangulate`
