+++ crates/fj-kernel/src/algorithms/triangulate/delaunay.rs
@@ -1,14 +1,32 @@
+use std::collections::BTreeSet;
+
 use fj_math::{Point, Scalar, Triangle, Winding};
 use spade::HasPosition;
 
-use crate::objects::Handedness;
+use crate::{algorithms::approx::cycle::CycleApprox, objects::Handedness};
 
 /// Create a Delaunay triangulation of all points
 pub fn triangulate(
-    points: Vec<TriangulationPoint>,
+    cycles: impl IntoIterator<Item = CycleApprox>,
     coord_handedness: Handedness,
 ) -> Vec<[TriangulationPoint; 3]> {
     use spade::Triangulation as _;
+
+    let points: Vec<_> = {
+        let mut points = BTreeSet::new();
+
+        for cycle_approx in cycles {
+            points.extend(cycle_approx.points());
+        }
+
+        points
+            .into_iter()
+            .map(|point| TriangulationPoint {
+                point_surface: point.local_form,
+                point_global: point.global_form,
+            })
+            .collect()
+    };
 
     let triangulation =
         spade::ConstrainedDelaunayTriangulation::<_>::bulk_load(points)
